requirements.txt
streamlit==1.25.0
transformers==4.34.0
diffusers>=0.18.0
torch==2.0.1
torchvision==0.15.2
Pillow>=9.5,<10

